@@ -5,79 +5,77 @@
   "angular-cli": {},
   "scripts": {
     "start": "ng serve --host 0.0.0.0",
-    "serve:prod": "ng serve -prod --sourcemap --host 0.0.0.0",
-    "lint": "tslint \"src/**/*.ts\"",
+    "serve:prod": "ng serve --prod --sourcemap --host 0.0.0.0",
+    "lint": "ng lint",
     "test": "ng test",
-    "pree2e": "webdriver-manager update",
-    "e2e": "protractor",
+    "e2e": "ng e2e",
     "prefixcss": "postcss -s postcss-scss -u autoprefixer --autoprefixer.browsers \"last 2 versions, iOS > 7, > 1% in US\" -r src/**/*.scss",
     "clone:index": "cp-cli dist/index.html dist/200.html",
     "surge": "surge -p dist -d trunkovich-paycal.surge.sh",
-    "build:prod": "ng build -prod --sourcemap",
+    "build:prod": "ng build --prod --sourcemap",
     "deploy": "npm run build:prod && npm run clone:index && npm run surge"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.*",
-    "@angular/cdk": "^5.1.1",
-    "@angular/common": "5.2.*",
-    "@angular/compiler": "5.2.*",
-    "@angular/core": "5.2.*",
-    "@angular/forms": "5.2.*",
-    "@angular/http": "5.2.*",
-    "@angular/material": "^5.1.1",
-    "@angular/platform-browser": "5.2.*",
-    "@angular/platform-browser-dynamic": "5.2.*",
-    "@angular/router": "5.2.*",
+    "@angular/animations": "~6.0.0",
+    "@angular/cdk": "~6.0.0",
+    "@angular/common": "~6.0.0",
+    "@angular/compiler": "~6.0.0",
+    "@angular/core": "~6.0.0",
+    "@angular/forms": "~6.0.0",
+    "@angular/http": "~6.0.0",
+    "@angular/material": "~6.0.0",
+    "@angular/platform-browser": "~6.0.0",
+    "@angular/platform-browser-dynamic": "~6.0.0",
+    "@angular/router": "~6.0.0",
     "@ngrx/core": "^1.2.0",
-    "@ngrx/effects": "^5.0.1",
-    "@ngrx/store": "^5.0.0",
-    "@ngrx/store-devtools": "^5.0.1",
+    "@ngrx/effects": "~6.0.0",
+    "@ngrx/store": "~6.0.0",
+    "@ngrx/store-devtools": "~6.0.0",
     "angular2-moment": "1.3.3",
     "angular2-text-mask": "2.1.0",
-    "angulartics2": "^5.0.1",
+    "angulartics2": "~6.2.0",
     "blueimp-canvas-to-blob": "3.6.0",
-    "core-js": "^2.5.1",
+    "core-js": "^2.5.4",
     "cropperjs": "1.0.0-beta.1",
     "intl": "1.2.5",
     "lodash": "4.17.4",
     "mobile-detect": "1.3.5",
     "ngrx-store-freeze": "^0.2.0",
-    "ngx-localstorage": "^1.2.0",
     "normalize.css": "5.0.0",
-    "proxy-polyfill": "0.1.6",
+    "proxy-polyfill": "~0.3.0",
     "raven-js": "^3.16.1",
     "reselect": "2.5.4",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.0.0",
     "text-mask-addons": "3.1.0",
     "ts-helpers": "1.1.1",
-    "zone.js": "^0.8.19"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.6",
-    "@angular/compiler-cli": "5.2.*",
+    "@angular-devkit/build-angular": "^0.6.8",
+    "@angular/cli": "~6.0.0",
+    "@angular/compiler-cli": "~6.0.0",
     "@types/cropperjs": "0.0.23",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
     "@types/lodash": "^4.14.63",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
+    "@types/node": "~8.9.4",
     "autoprefixer": "6.6.1",
-    "codelyzer": "~3.1.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "codelyzer": "~4.2.1",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~1.7.1",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
     "postcss": "5.2.10",
     "postcss-cli": "2.6.0",
     "postcss-scss": "0.4.0",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
+    "protractor": "~5.3.0",
+    "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3",
+    "typescript": "~2.7.2",
     "webdriver-manager": "11.1.1"
   }
 }
